Feature: Implemented Error handling

+++ src/main/java/com/juls/labs/websorting/service/impl/EventServiceImpl.java
@@ -1,5 +1,6 @@
 package com.juls.labs.websorting.service.impl;
 
+import com.juls.labs.websorting.exception.EventNotFoundException;
 import com.juls.labs.websorting.model.Event;
 import com.juls.labs.websorting.repository.EventRepository;
 import com.juls.labs.websorting.repository.impl.EventRepositoryImpl;
@@ -31,8 +32,9 @@ public class EventServiceImpl implements EventService {
     }
 
     @Override
-    public Optional<Event> getEventById(Long eventId) {
-        return this.eventRepository.findById(eventId);
+    public Optional<Event> getEventById(Long eventId) throws EventNotFoundException {
+        return Optional.ofNullable(this.eventRepository.findById(eventId)
+                .orElseThrow(() -> new EventNotFoundException("Event with Id does not exist: " + eventId)));
     }
 
 
@@ -42,19 +44,30 @@ public class EventServiceImpl implements EventService {
         return this.eventRepository.findAll();
     }
 
+    private boolean eventExist(Long eventId) throws EventNotFoundException {
+        return this.getEventById(eventId).isPresent();
+    }
+
     @Override
-    public Event updateEvent(Event event) {
+    public Event updateEvent(Event event) throws EventNotFoundException {
+        if (eventExist(event.getEventId())){
+            throw new EventNotFoundException("Event does not exist");
+        }
         return this.eventRepository.update(event);
     }
 
     @Override
-    public void deleteEvent(Long eventId) {
+    public void deleteEvent(Long eventId) throws EventNotFoundException {
+        if (eventExist(eventId)){
+            throw new EventNotFoundException("Event does not exist");
+        }
         this.eventRepository.delete(eventId);
     }
 
     @Override
-    public Optional<Event> searchEvent(String eventName) {
-        return this.eventRepository.getEventByName(eventName);
+    public Optional<Event> searchEvent(String eventName) throws EventNotFoundException {
+        return Optional.ofNullable(this.eventRepository.getEventByName(eventName))
+                .orElseThrow(() -> new EventNotFoundException("Event Not found"));
     }
 
     @Override
@@ -64,6 +77,9 @@ public class EventServiceImpl implements EventService {
 
     @Override
     public List<Event> getEventByDate(LocalDate eventDate) {
+        if(eventDate == null){
+            throw new IllegalArgumentException("Specify a date value");
+        }
         return this.eventRepository.getEventByDate(eventDate);
     }
